+++ CHANGES.md
@@ -12,6 +12,7 @@ Changes:
 * Most of your blog's configuration is now stored in MongoDB and can be modified from the new "Configuration" page in your the admin panel.
 * When logged in as an admin, you will be shown a small admin actions panel in the upper right corner of your browser, allowing you to quickly edit posts, jump to the admin section, or log out.
 * Schnitzelpress now has a light-weight, custom-built asset pipeline that serves all Javascripts and Stylesheets as one single file each, compressed and ready for hardcore caching.
+* When running Schnitzelpress locally (aka: development mode), you can use a simple developer-only login provider to log into your blog for testing purposes.
 * Post with dates now use double-digit days and months in their canonical URLs. (Your existing posts will forward to the new canonical URLs automatically.)
 * Various performance improvements.